# skills/bkd/references/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BKD API",
    "description": "Kanban board for managing AI coding agents. Issues are assigned to CLI-based AI engines (Claude Code, Codex, Gemini CLI) that execute autonomously.",
    "version": "dev",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "channel": {
            "type": "string",
            "enum": [
              "webhook",
              "telegram"
            ]
          },
          "url": {
            "type": "string"
          },
          "secret": {
            "type": [
              "string",
              "null"
            ],
            "description": "Masked in responses"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "channel",
          "url",
          "secret",
          "events",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateWebhook": {
        "type": "object",
        "properties": {
          "channel": {
            "type": "string",
            "enum": [
              "webhook",
              "telegram"
            ],
            "default": "webhook"
          },
          "url": {
            "type": "string",
            "minLength": 1,
            "description": "Webhook URL or Telegram chat ID"
          },
          "secret": {
            "type": "string",
            "maxLength": 256,
            "description": "HMAC secret or Telegram bot token"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "issue.created",
                "issue.updated",
                "issue.deleted",
                "issue.status.todo",
                "issue.status.working",
                "issue.status.review",
                "issue.status.done",
                "session.started",
                "session.completed",
                "session.failed"
              ]
            },
            "minItems": 1
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "url",
          "events"
        ]
      },
      "UpdateWebhook": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1
          },
          "secret": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 256
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "issue.created",
                "issue.updated",
                "issue.deleted",
                "issue.status.todo",
                "issue.status.working",
                "issue.status.review",
                "issue.status.done",
                "session.started",
                "session.completed",
                "session.failed"
              ]
            },
            "minItems": 1
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/settings/webhooks": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "List webhooks",
        "operationId": "listWebhooks",
        "responses": {
          "200": {
            "description": "Webhook list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Webhook"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Create webhook",
        "operationId": "createWebhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhook"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created webhook",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Webhook"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/webhooks/{webhookId}": {
      "patch": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Update webhook",
        "operationId": "updateWebhook",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "webhookId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated webhook",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Webhook"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Webhook not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Soft-delete webhook",
        "operationId": "deleteWebhook",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "webhookId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Webhook not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/webhooks/{webhookId}/test": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Send test webhook delivery",
        "operationId": "testWebhook",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "webhookId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "sent": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "sent"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Webhook not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}